# mycelix-workspace/mycelix-music/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mycelix Music API
  description: |
    Decentralized music streaming platform API.

    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Tokens are obtained by signing a message with your Ethereum wallet.

    ## Rate Limiting
    - Anonymous: 100 requests/minute
    - Authenticated: 500 requests/minute
    - Premium: 1000 requests/minute

    Rate limit headers are included in all responses.
  version: 1.0.0
  contact:
    name: Mycelix Support
    email: support@mycelix.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mycelix.io/v1
    description: Production
  - url: https://api.staging.mycelix.io/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Local development
...
paths:
...
  /songs:
    get:
      tags: [Songs]
      summary: List songs
      description: Get a paginated list of songs with optional filters
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: genre
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [plays, releaseDate, title]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongList'
  /songs/{id}:
    get:
      tags: [Songs]
      summary: Get song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Song details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
...
    Song:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistSummary'
        coverArt:
          type: string
          format: uri
        duration:
          type: integer
          description: Duration in seconds
        genre:
          type: string
        tags:
          type: array
          items:
            type: string
        plays:
          type: integer
        likes:
          type: integer
        explicit:
          type: boolean
        releaseDate:
          type: string
          format: date
    SongList:
      type: object
      properties:
        songs:
          type: array
          items:
            $ref: '#/components/schemas/Song'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ArtistSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        verified:
          type: boolean
...